+++ CHANGES.rst
@@ -20,6 +20,9 @@ Bug Fixes
 - Mark5B reader more robust against headers not being parsed correctly
   in ``Mark5BFileReader.find_header``. [#275]
 
+- All stream readers now have a proper ``dtype`` attribute, not a
+  corresponding ``np.float32`` or ``np.complex64``. [#280]
+
 1.2 (2018-07-27)
 ================